Bug fix for unsunk ships with no moves
Felt like this should've been found much sooner.
This situation happens when a ship gets hit (not through collateral of another ship), then it's direction is found, but it hits another ship before it's sunk. This leaves it with no moves and it's most likely to be picked out of the two ships.